# perf(detect-libc): faster musl check by looking for ldd file