--- a/benchmark/call-isNonGlibcLinuxSync.js
+++ b/benchmark/call-isNonGlibcLinuxSync.js
@@ -1,0 +1,7 @@
+const performance = require('perf_hooks').performance;
+const libc = require('..');
+
+const now = performance.now();
+libc.isNonGlibcLinuxSync();
+
+console.log(`[isNonGlibcLinux] Time Spent ${performance.now() - now}ms`);