--- a/lib/detect-libc.js
+++ b/lib/detect-libc.js
@@ -5,6 +5,10 @@
 
 const childProcess = require('child_process');
 const { isLinux, getReport } = require('./process');
+const { LDD_PATH, readFile, readFileSync } = require('./filesystem');
+
+let cachedFamilyFilesystem;
+let cachedVersionFilesystem;
 
 const command = 'getconf GNU_LIBC_VERSION 2>&1 || true; ldd --version 2>&1 || true';
 let commandOut = '';
@@ -40,11 +44,29 @@ const safeCommandSync = () => {
 const GLIBC = 'glibc';
 
 /**
+ * A Regexp constant to get the GLIBC Version.
+ * @type {string}
+ */
+const RE_GLIBC_VERSION = /GLIBC\s(\d+\.\d+)/;
+
+/**
  * A String constant containing the value `musl`.
  * @type {string}
  * @public
  */
 const MUSL = 'musl';
+
+/**
+ * This string is used to find if the {@link LDD_PATH} is GLIBC
+ * @type {string}
+ */
+const GLIBC_ON_LDD = GLIBC.toUpperCase();
+
+/**
+ * This string is used to find if the {@link LDD_PATH} is musl
+ * @type {string}
+ */
+const MUSL_ON_LDD = MUSL.toLowerCase();
 
 const isFileMusl = (f) => f.includes('libc.musl-') || f.includes('ld-musl-');
 
@@ -72,6 +94,40 @@ const familyFromCommand = (out) => {
   return null;
 };
 
+const getFamilyFromLddContent = (content) => {
+  if (content.includes(MUSL_ON_LDD)) {
+    return MUSL;
+  }
+  if (content.includes(GLIBC_ON_LDD)) {
+    return GLIBC;
+  }
+  return null;
+};
+
+const familyFromFilesystem = async () => {
+  if (cachedFamilyFilesystem !== undefined) {
+    return cachedFamilyFilesystem;
+  }
+  cachedFamilyFilesystem = null;
+  try {
+    const lddContent = await readFile(LDD_PATH);
+    cachedFamilyFilesystem = getFamilyFromLddContent(lddContent);
+  } catch (e) {}
+  return cachedFamilyFilesystem;
+};
+
+const familyFromFilesystemSync = () => {
+  if (cachedFamilyFilesystem !== undefined) {
+    return cachedFamilyFilesystem;
+  }
+  cachedFamilyFilesystem = null;
+  try {
+    const lddContent = readFileSync(LDD_PATH);
+    cachedFamilyFilesystem = getFamilyFromLddContent(lddContent);
+  } catch (e) {}
+  return cachedFamilyFilesystem;
+};
+
 /**
  * Resolves with the libc family when it can be determined, `null` otherwise.
  * @returns {Promise<?string>}
@@ -79,7 +135,10 @@ const familyFromCommand = (out) => {
 const family = async () => {
   let family = null;
   if (isLinux()) {
-    family = familyFromReport();
+    family = await familyFromFilesystem();
+    if (!family) {
+      family = familyFromReport();
+    }
     if (!family) {
       const out = await safeCommand();
       family = familyFromCommand(out);
@@ -95,7 +154,10 @@ const family = async () => {
 const familySync = () => {
   let family = null;
   if (isLinux()) {
-    family = familyFromReport();
+    family = familyFromFilesystemSync();
+    if (!family) {
+      family = familyFromReport();
+    }
     if (!family) {
       const out = safeCommandSync();
       family = familyFromCommand(out);
@@ -115,6 +177,36 @@ const isNonGlibcLinux = async () => isLinux() && await family() !== GLIBC;
  * @returns {boolean}
  */
 const isNonGlibcLinuxSync = () => isLinux() && familySync() !== GLIBC;
+
+const versionFromFilesystem = async () => {
+  if (cachedVersionFilesystem !== undefined) {
+    return cachedVersionFilesystem;
+  }
+  cachedVersionFilesystem = null;
+  try {
+    const lddContent = await readFile(LDD_PATH);
+    const versionMatch = lddContent.match(RE_GLIBC_VERSION);
+    if (versionMatch) {
+      cachedVersionFilesystem = versionMatch[1];
+    }
+  } catch (e) {}
+  return cachedVersionFilesystem;
+};
+
+const versionFromFilesystemSync = () => {
+  if (cachedVersionFilesystem !== undefined) {
+    return cachedVersionFilesystem;
+  }
+  cachedVersionFilesystem = null;
+  try {
+    const lddContent = readFileSync(LDD_PATH);
+    const versionMatch = lddContent.match(RE_GLIBC_VERSION);
+    if (versionMatch) {
+      cachedVersionFilesystem = versionMatch[1];
+    }
+  } catch (e) {}
+  return cachedVersionFilesystem;
+};
 
 const versionFromReport = () => {
   const report = getReport();
@@ -144,7 +236,10 @@ const versionFromCommand = (out) => {
 const version = async () => {
   let version = null;
   if (isLinux()) {
-    version = versionFromReport();
+    version = await versionFromFilesystem();
+    if (!version) {
+      version = versionFromReport();
+    }
     if (!version) {
       const out = await safeCommand();
       version = versionFromCommand(out);
@@ -160,7 +255,10 @@ const version = async () => {
 const versionSync = () => {
   let version = null;
   if (isLinux()) {
-    version = versionFromReport();
+    version = versionFromFilesystemSync();
+    if (!version) {
+      version = versionFromReport();
+    }
     if (!version) {
       const out = safeCommandSync();
       version = versionFromCommand(out);