--- a/package.json
+++ b/package.json
@@ -8,7 +8,9 @@
     "index.d.ts"
   ],
   "scripts": {
-    "test": "semistandard && nyc --reporter=lcov --check-coverage --branches=100 ava test/unit.js"
+    "test": "semistandard && nyc --reporter=lcov --check-coverage --branches=100 ava test/unit.js",
+    "bench": "node benchmark/detect-libc",
+    "bench:calls": "node benchmark/call-familySync.js && sleep 1 && node benchmark/call-isNonGlibcLinuxSync.js && sleep 1 && node benchmark/call-versionSync.js"
   },
   "repository": {
     "type": "git",
@@ -26,6 +28,7 @@
   "license": "Apache-2.0",
   "devDependencies": {
     "ava": "^2.4.0",
+    "benchmark": "^2.1.4",
     "nyc": "^15.1.0",
     "proxyquire": "^2.1.3",
     "semistandard": "^14.2.3"