… test/fixtexture-file.txt