Closes #18 

Now, we use by default the comparison of check if `ldd` file exists.

## Performance

Before:

```
[family] Time Spent 30.49158699810505ms
[isNonGlibcLinux] Time Spent 29.517467997968197ms
[versionSync] Time Spent 26.344385005533695ms
```
> npm run bench:calls

Now:

```
[family] Time Spent 0.15008499845862389ms
[isNonGlibcLinux] Time Spent 0.1661360003054142ms
[versionSync] Time Spent 0.184239000082016ms
```
> npm run bench:calls