@H4ad 2023-06-16
@lovell I don't know if `perf` is appropriated to describe this change since it could be a breaking change, so let me know if you want me to change to other message.