@lovell 2023-06-16
Could we use this same approach for a potential `versionFromFilesystem()`? For example, the `/usr/bin/ldd` file on some glibc-based Linux will contain the string literal `"GLIBC X.Y"`.