@H4ad 2023-06-16
In theory, we can:

![image](https://github.com/lovell/detect-libc/assets/12551007/97bed5a7-1b62-41c9-a595-795b9093448f)
> This is the content of /usr/bin/ldd

We just need to build a faster parser to extract this information.

Do you want to include the change on version in this PR too?