@H4ad 2023-06-17
@lovell How did you test linux musl? I'm trying to understand what is the output of `cat /usr/bin/ldd` on musl