@lovell 2023-06-17
Although we can probably determine the family from the `/usr/bin/ldd` file on musl-based Linux, we are unlikely to be able to determine the version.

```sh
$ docker run -it --rm alpine:3.18 cat /usr/bin/ldd
#!/bin/sh
exec /lib/ld-musl-x86_64.so.1 --list "$@"
```