@lovell 2023-07-18
v2.0.2 now available, tudo bom, obrigado!